change to ansible-lint version prior to automagic ansible detection
ansible-lint will helpfully also pull in `ansible` as a module
if there is an 'ansible' in the same venv.
However, this isn't a general solution for running other binaries from the same venv.
 I would rather force using more robust features of PEX or the toolchain
so that it's easier to use this as a base for other plugins
or rules within this plugin