use `--venv prepend` for PEX to get multiple binaries working
This option isn't plumbed with the VenvPexRequest,
so I just use the PexRequest with additional args which do the same.
VenvPexRequest mostly just adds the `--venv` argument to the PEX creation,
so using the PexRequest in this way is mostly equivalent anyhow.

`--venv prepend` prepends the binaries to the PATH, so they're discoverable