tell ansible-playbook to use installed collections
using the envvar ANSIBLE_COLLECTIONS_PATHS,
we can use the installed collections
regardless of the location of the playbook
and without relying on the local collections path.
This also means that we don't have to relocate playbooks.
Keeping playbooks in their positions in the repo
makes relative references cleaner.
For example, importing from /envs/prod/vars.yml